+++ react-client/package.json
@@ -97,7 +97,7 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-  "proxy": "http://localhost:3021",
+  "proxy": "http://localhost:3022",
   "eslintConfig": {
     "extends": "react-app"
   },
