+++ react-client/src/components/LiveChat/index.js
@@ -1,0 +1,3 @@
+import INFO from './info/index.js'
+
+export { INFO }
