+++ react-client/src/routes.js
@@ -47,7 +47,7 @@ import FeedPosts from './containers/rssIntegration/feedPosts'
 import NewsPosts from './containers/newsIntegration/newsPosts'
 import MessagesContainer from './containers/autoposting/messagesContainer'
 import ItemSettings from './containers/autoposting/itemSettings'
-import LiveChat from './containers/liveChat/livechat'
+import LiveChat from './containers/liveChat/liveChat'
 import Settings from './containers/settings/settings'
 import CreateSurvey from './containers/templates/createSurvey'
 import CreateTemplatePoll from './containers/templates/createPoll'
